pyproject.toml
[project]
name = "coact_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "dotenv>=0.9.9",
    "hud-python>=0.2.1",
    "ipykernel>=6.29.5",
    "requests>=2.32.4",
    "supabase>=2.15.3",
    "openai",
    "typing-extensions",
    "autogen-agentchat>=0.7.2",
    "autogen>=0.9.8.post1",
]
